change current evaluation to comparison and create new evaluation that does not train the model again but just runs preexisting model on test set
