+++ sudoku_digitalisation/main.py
@@ -2,7 +2,8 @@ from sudoku_digitalisation.models.SVM import SVM
 from sudoku_digitalisation.features.sudoku_preprocessing import get_preprocessor
 from sudoku_digitalisation.scripts.run_training import get_model
 from sudoku_digitalisation.scripts.run_prediction import make_prediction
-from sudoku_digitalisation.scripts.run_evaluation import compare_cnn_svm
+from sudoku_digitalisation.scripts.run_comparison import compare_cnn_svm
+from sudoku_digitalisation.scripts.run_evaluation import evaluate_model
 
 # Run: "python -m sudoku_digitalisation.main" if you have ModuleNotFoundError
 if __name__ == "__main__":
@@ -11,8 +12,9 @@ if __name__ == "__main__":
     TRAIN_CNN = False          # change this to True if you want to train the CNN, otherwise it is loaded
     TRAIN_SVM = False           # change this to True if you want to train the SVM (baseline), otherwise it is loaded
     GET_SVM = False             # if this is set to False, the SVM is not gotten
-    EVALUATE = False             # change to true if you want to run the evaluation
-    PREDICT = True            # change to true if you want to run the prediction
+    COMPARISON = False          # change to True if you want to compare our CNN model to the SVM baseline
+    EVALUATE = True             # change to true if you want to run the evaluation
+    PREDICT = False            # change to true if you want to run the prediction
 
     CNN_NAME = '32_64_128'     # name for how you save and load CNN
     SVM_NAME = 'default'        # name for how you save and load SVM
@@ -29,8 +31,11 @@ if __name__ == "__main__":
     # perform k-fold cross validation of CNN and SVM and compare results
     # this is based off of the implementation of the CNN and SVM in the files, not the loaded models
     # (because they need to be trained anyways for k-fold)
-    if EVALUATE:
+    if COMPARISON:
         compare_cnn_svm(preprocessor, k=5)
+
+    if EVALUATE:
+        evaluate_model(cnn, preprocessor, TRAIN_CNN)
 
     # prediction if we want but that would be more for the API
     if PREDICT:
