+++ sudoku_digitalisation/scripts/run_evaluation.py
@@ -1,100 +1,64 @@
 import numpy as np
 import matplotlib.pyplot as plt
-from sklearn.model_selection import KFold
-from typing import Tuple, Dict, List
-from sklearn.metrics import ConfusionMatrixDisplay
-from sudoku_digitalisation.features.sudoku_preprocessing import DatasetPreprocessor
+from typing import Dict, Tuple, Union, Any, List
 from sudoku_digitalisation.models.CNN import CNN
-from sudoku_digitalisation.models.SVM import SVM
+from sudoku_digitalisation.features.sudoku_preprocessing import DatasetPreprocessor
+from sklearn.metrics import ConfusionMatrixDisplay
 
+def get_test(preprocessor: DatasetPreprocessor):
+    '''
+    Get the test set from the preprocessor.
+    '''
+    digit_dataset = preprocessor.handler.datasets['digits']
 
-def kfold_evaluation(
-        model_type: str,
-        preprocessor: DatasetPreprocessor,
-        k: int,
-        val_split: float = 0.2
-        ) -> Tuple[Dict[str, float], Dict[str, float]]:
-    """
-    Perform k-fold cross-validation and return mean and std of metrics.
-    """
-    kf = KFold(n_splits=k, shuffle=True)
-    cm_total = None
-    metrics_folds = []
+    X_test = digit_dataset['test']['image']
+    y_test = digit_dataset['test']['label']
 
-    train = preprocessor.handler.datasets['digits']['train']
-    val = preprocessor.handler.datasets['digits']['validation']
-    test = preprocessor.handler.datasets['digits']['test']
+    return X_test, y_test
 
-    images = np.array(train['image'] + val['image'] + test['image'])
-    labels = np.array(train['label'] + val['label'] + test['label'])
+def evaluate_model(cnn: CNN, preprocessor: DatasetPreprocessor, trained: bool):
+    '''
+    Evaluates the model's accuracy, precision, recall and F1
+    '''
+    X_test, y_test = get_test(preprocessor)
 
-    dim = preprocessor.cropper.output_size // 9
+    cm, info, history = cnn.evaluate(X_test, y_test)
 
-    for train_index, test_index in kf.split(images):
-        X_train_full, X_test = images[train_index], images[test_index]
-        y_train_full, y_test = labels[train_index], labels[test_index]
-
-        if model_type == 'cnn':
-            val_size = int(len(X_train_full) * val_split)
-            X_val = X_train_full[:val_size]
-            y_val = y_train_full[:val_size]
-            X_train = X_train_full[val_size:]
-            y_train = y_train_full[val_size:]
-
-            model = CNN(input_shape=(dim, dim, 1), num_classes=10)
-            model.train(X_train, y_train, X_val, y_val)
-        elif model_type == 'svm':
-            model = SVM(input_shape=(dim, dim))
-            model.train(X_train_full, y_train_full)
-        else:
-            raise ValueError(f"Unknown model_type '{model_type}'. Expected 'cnn' or 'svm'.")
-
-        cm, metrics = model.evaluate(X_test, y_test)
-        if cm_total is None:
-            cm_total = cm
-        else:
-            cm_total += cm
-        metrics_folds.append(metrics)
-
-    return cm_total, compute_mean_std(metrics_folds)   
-
-
-def compute_mean_std(results: List[Dict[str, int]]) -> Dict[str, Dict[str, float]]:
-    metric_names = results[0].keys()
-
-    summary = {}
-    for metric in metric_names:
-        values = np.array([fold[metric] for fold in results])
-        summary[metric] = {
-            "mean": float(np.mean(values)),
-            "std": float(np.std(values))
-        }
-    return summary
-
-
-def compare_cnn_svm(
-        preprocessor: DatasetPreprocessor,
-        k: int
-        ) -> None:
-    cnn_cm, cnn_metrics = kfold_evaluation('cnn', preprocessor, k)
-    svm_cm, svm_metrics = kfold_evaluation('svm', preprocessor, k)
-
-    print(f"{'Metric':<15} | {'CNN':^15} | {'SVM':^15}")
-    print("-" * 50)
-    for metric in cnn_metrics.keys():
-        cnn_mean = cnn_metrics[metric]["mean"]
-        cnn_std = cnn_metrics[metric]["std"]
-        svm_mean = svm_metrics[metric]["mean"]
-        svm_std = svm_metrics[metric]["std"]
-        print(f"{metric:<15} | {cnn_mean:.4f} ± {cnn_std:.4f} | {svm_mean:.4f} ± {svm_std:.4f}")
-    print()
-
-    show_cm(cnn_cm)
-    show_cm(svm_cm)
-    
-
-def show_cm(cm):
-    disp = ConfusionMatrixDisplay(confusion_matrix=cm)
+    # Confusion matrix
+    disp = ConfusionMatrixDisplay(confusion_matrix=cm, display_labels=range(len(cm)))
     disp.plot(cmap='viridis', values_format='d')
-    plt.title("Confusion Matrix")
+    plt.title('Confusion Matrix')
     plt.show()
+
+    # Per class accuracy
+    class_accuracies = cm.diagonal() / cm.sum(axis=1)
+    for i, acc in enumerate(class_accuracies):
+        print(f"Accuracy for Class {i}: {acc:.4f}")
+    print(f"Accuracy for test set: {info['test accuracy']:.4f}")
+
+    # Per class precision, recall and F1
+    print(f"The class precision is: {info['precision macro']:.4f}")
+    print(f"The recall is: {info['recall macro']:.4f}")
+    print(f"The f1 score is: {info['f1 macro']:.4f}")
+
+    # accuracy on sudokus
+    print(f"accuracy of fully correct sudokus: {info['sudoku accuracy']}")
+
+    if trained:
+        # Accuracy plot over time
+        plt.plot(history.history['accuracy'], label='Training Accuracy')
+        plt.plot(history.history['val_accuracy'], label='Validation Accuracy')
+        plt.xlabel('Epochs')
+        plt.ylabel('Accuracy')
+        plt.legend()
+        plt.title('Training vs Validation Accuracy')
+        plt.show()
+
+        # Loss plot over time
+        plt.plot(history.history['loss'], label='Training Loss')
+        plt.plot(history.history['val_loss'], label='Validation Loss')
+        plt.xlabel('Epochs')
+        plt.ylabel('Loss')
+        plt.legend()
+        plt.title('Training vs Validation Loss')
+        plt.show()
